Clean up code audit findings

+++ services/commerce/internal/api/openapi_gen.go
@@ -367,13 +367,14 @@ type CustomerOrderSummary struct {
 	Quantity      int                `json:"quantity"`
 	StartsAt      *time.Time         `json:"starts_at"`
 
-	// TotalAmount Integer minor units (ADR-001). Never a float, never divided here.
-	TotalAmount int `json:"total_amount"`
+	// TotalAmount The composed order total in integer minor units (ADR-001). It may exceed a JavaScript Number's safe-integer range; number-based clients must reject a value they cannot represent exactly.
+	TotalAmount int64 `json:"total_amount"`
 }
 
 // CustomerPrincipal A signed-in buyer. `email` is the address in the spelling the buyer registered with, for display; every lookup goes through the normalized form, which is never exposed.
 type CustomerPrincipal struct {
 	// CustomerAssertion A commerce-signed, expiring proof that the holder authenticated as this customer, presented on checkout in `X-Customer-Assertion` (TKT-221, ADR-049).
+	// Version 1 wire form is `v1.<non-nil customer UUID>.<positive Unix expiry>.<43-character base64url HMAC>`. The expiry fits a JavaScript safe integer.
 	// This is a BEARER CREDENTIAL: anyone holding it can attribute a checkout to this customer until it expires. It is returned only to a caller who has just proven the password, and the storefront keeps it server-side inside its in-process session — never in a cookie, a rendered prop, or a log. Its lifetime is deliberately equal to that session's, so "the session is alive" and "the assertion is valid" cannot disagree and strand a buyer at the payment button.
 	CustomerAssertion string             `json:"customer_assertion"`
 	CustomerId        openapi_types.UUID `json:"customer_id"`
@@ -584,6 +585,7 @@ type PasswordResetResult struct {
 
 // Refund defines model for Refund.
 type Refund struct {
+	// Amount This refund's integer minor-unit amount: the persisted face-value unit amount multiplied by the refunded quantity. Passed-on fees are not part of this value. The product can exceed a JavaScript Number's safe-integer range, so number-based clients must reject a value they cannot represent exactly.
 	Amount           int64              `json:"amount"`
 	CapacityReturned bool               `json:"capacity_returned"`
 	Currency         string             `json:"currency"`
@@ -591,10 +593,12 @@ type Refund struct {
 	Quantity         int                `json:"quantity"`
 	RefundId         openapi_types.UUID `json:"refund_id"`
 	RefundStatus     RefundRefundStatus `json:"refund_status"`
-	RefundedAmount   int64              `json:"refunded_amount"`
-	RefundedQuantity int                `json:"refunded_quantity"`
-	Replay           bool               `json:"replay"`
-	TicketsVoided    bool               `json:"tickets_voided"`
+
+	// RefundedAmount The cumulative sum of face-value refund amounts in integer minor units. It can exceed a JavaScript Number's safe-integer range, so number-based clients must reject a value they cannot represent exactly.
+	RefundedAmount   int64 `json:"refunded_amount"`
+	RefundedQuantity int   `json:"refunded_quantity"`
+	Replay           bool  `json:"replay"`
+	TicketsVoided    bool  `json:"tickets_voided"`
 }
 
 // RefundRefundStatus defines model for Refund.RefundStatus.
@@ -610,17 +614,19 @@ type RefundCreate struct {
 
 // Reservation defines model for Reservation.
 type Reservation struct {
+	// Amount The composed order total in integer minor units. It may exceed a JavaScript Number's safe-integer range because individually bounded unit prices and fees are multiplied by quantity. Number-based clients must reject a value they cannot represent exactly.
 	Amount    int64              `json:"amount"`
 	BuyerId   openapi_types.UUID `json:"buyer_id"`
 	Currency  string             `json:"currency"`
 	ExpiresAt time.Time          `json:"expires_at"`
 
 	// FaceValue The rule-resolved price times quantity, before any fee (TKT-215). `amount` minus `face_value` is exactly the passed-on fee total.
-	// OPTIONAL, like `seats` and for the same reason: every reservation created before this field existed lacks it, and the response validator fails closed (ADR-028), so requiring it would turn those valid responses into runtime 500s. It is absent on a reservation with no fee snapshot -- a pre-TKT-215 row or a staff-created one -- and absence means "this sale had no fee concept", which is a different fact from "its fees totalled nothing".
+	// OPTIONAL, like `seats` and for the same reason: every reservation created before this field existed lacks it, and the response validator fails closed (ADR-028), so requiring it would turn those valid responses into runtime 500s. It is absent on a reservation with no fee snapshot -- a pre-TKT-215 row or a staff-created one -- and absence means "this sale had no fee concept", which is a different fact from "its fees totalled nothing". Multiplication by quantity can put this value beyond a JavaScript Number's safe-integer range.
 	FaceValue *int64 `json:"face_value,omitempty"`
 
 	// FeeBreakdown One entry per fee code that applied, INCLUDING any whose computed amount is zero (ADR-046 §2) -- a code that vanished as a function of price would leave settlement with a payee that is sometimes owed nothing and sometimes absent. A fee code that was considered but had no live rule produces no entry at all.
 	FeeBreakdown *[]struct {
+		// Amount The fee after applying its basis to the reservation quantity. This composed amount may exceed a JavaScript Number's safe-integer range.
 		Amount    int64                            `json:"amount"`
 		Basis     ReservationFeeBreakdownBasis     `json:"basis"`
 		Currency  string                           `json:"currency"`
@@ -629,7 +635,7 @@ type Reservation struct {
 	} `json:"fee_breakdown,omitempty"`
 	HoldId openapi_types.UUID `json:"hold_id"`
 
-	// PassedOnFees The part of `amount` that is fees the buyer pays. Absorbed fees are NOT here and never reach `amount`: they are borne by the organizer out of the face value, and charging them to the buyer would be charging for the organizer's cost. They are recorded in the persisted snapshot because TKT-217 still owes them to a payee.
+	// PassedOnFees The part of `amount` that is fees the buyer pays. Absorbed fees are NOT here and never reach `amount`: they are borne by the organizer out of the face value, and charging them to the buyer would be charging for the organizer's cost. They are recorded in the persisted snapshot because TKT-217 still owes them to a payee. Their sum can exceed a JavaScript Number's safe-integer range.
 	PassedOnFees  *int64             `json:"passed_on_fees,omitempty"`
 	ReservationId openapi_types.UUID `json:"reservation_id"`
